requirements.txt
aiohttp==3.8.6
aiohttp-cors==0.7.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorful==0.5.5
cryptography==41.0.4
distlib==0.3.7
filelock==3.12.4
flwr==1.5.0
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.12.0
google-auth==2.23.3
googleapis-common-protos==1.61.0
gpustat==1.1.1
grpcio==1.59.0
hydra-core==1.3.2
idna==3.4
iterators==0.0.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.2
numpy==1.26.1
nvidia-ml-py==12.535.108
omegaconf==2.3.0
opencensus==0.11.3
opencensus-context==0.1.3
packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
protobuf==3.20.3
psutil==5.9.6
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smart-open==6.4.0
sympy==1.12
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.21.0
wcwidth==0.2.8
yarl==1.9.2
